consolidate on caching dump worker class name
